New objects, new icons for delete/clearAll/Stretching/TabToPlace, app icon for all platform, fixed icon highlighting, first version of documentation
I added four new Objects:
- Hemisphere
- Tube
- Ring
-Prism
I changed the Icons for deleting an object, deleting all objects, stretching in different directions and tab to place.
The highlighting of the icon, that hides all indicators works.
Furthermore there is now an default app icon for all plattforms.
I added a new folder with the documentation, but we still have to insert the pictures.